+++ core/src/main/kotlin/com/gitlab/kordlib/core/entity/channel/DmChannel.kt
@@ -4,6 +4,7 @@ import com.gitlab.kordlib.core.Kord
 import com.gitlab.kordlib.core.cache.data.ChannelData
 import com.gitlab.kordlib.common.entity.Snowflake
 import com.gitlab.kordlib.core.behavior.MessageBehavior
+import com.gitlab.kordlib.core.behavior.UserBehavior
 import com.gitlab.kordlib.core.behavior.channel.ChannelBehavior
 import com.gitlab.kordlib.core.behavior.channel.MessageChannelBehavior
 import com.gitlab.kordlib.core.entity.User
@@ -22,6 +23,11 @@ data class DmChannel(override val data: ChannelData, override val kord: Kord) : 
      * The ids of the recipients of the channel.
      */
     val recipientIds: Set<Snowflake> get() = data.recipients.orEmpty().asSequence().map { Snowflake(it) }.toSet()
+
+    /**
+     * The behaviors of the recipients of the channel.
+     */
+    val recipientBehaviors: Set<UserBehavior> get() = recipientIds.map { UserBehavior(it, kord) }.toSet()
 
     /**
      * Requests to get the recipients of the channel.