+++ core/src/main/kotlin/com/gitlab/kordlib/core/entity/User.kt
@@ -1,6 +1,7 @@
 package com.gitlab.kordlib.core.entity
 
 import com.gitlab.kordlib.common.entity.Snowflake
+import com.gitlab.kordlib.common.entity.UserFlags
 import com.gitlab.kordlib.core.Kord
 import com.gitlab.kordlib.core.KordObject
 import com.gitlab.kordlib.core.behavior.UserBehavior
@@ -27,6 +28,11 @@ open class User(val data: UserData, override val kord: Kord) : UserBehavior {
      * The 4-digit code at the end of the user's discord tag.
      */
     val discriminator: String get() = data.discriminator
+
+    /**
+     * The flags on a user's account, if present.
+     */
+    val flags: UserFlags? get() = data.flags
 
     /**
      * The complete user tag.