+++ common/src/main/kotlin/com/gitlab/kordlib/common/entity/DiscordUser.kt
@@ -115,6 +115,7 @@ data class DiscordOptionallyMemberUser(
 enum class Premium(val code: Int) {
     /** The default code for unknown values. */
     Unknown(Int.MIN_VALUE),
+    None(0),
     NitroClassic(1),
     Nitro(2);