+++ core/src/main/kotlin/com/gitlab/kordlib/core/entity/User.kt
@@ -1,5 +1,6 @@
 package com.gitlab.kordlib.core.entity
 
+import com.gitlab.kordlib.common.entity.Premium
 import com.gitlab.kordlib.common.entity.Snowflake
 import com.gitlab.kordlib.common.entity.UserFlags
 import com.gitlab.kordlib.core.Kord
@@ -33,6 +34,11 @@ open class User(val data: UserData, override val kord: Kord) : UserBehavior {
      * The flags on a user's account, if present.
      */
     val flags: UserFlags? get() = data.flags
+
+    /**
+     * The type of Nitro subscription on a user's account, if present.
+     */
+    val premiumType: Premium? get() = data.premium
 
     /**
      * The complete user tag.