+++ CHANGELOG.md
@@ -10,7 +10,8 @@ through rest.
 * Added missing `MessageUpdateEvent#message`, `MessageUpdateEvent#channel`, `MessageUpdateEvent#getMessage`,
 `MessageUpdateEvent#getChannel`.
 * Added missing `PresenceUpdateEvent#member`, `PresenceUpdateEvent#guild`, `PresenceUpdateEvent#getUser`,
-`PresenceUpdateEvent#getMember`, `PresenceUpdateEvent#member`, `PresenceUpdateEvent#getGuild`
+`PresenceUpdateEvent#getMember`, `PresenceUpdateEvent#member`, `PresenceUpdateEvent#getGuild`.
+* Added `chunk_index` and `chunck_count` properties to `GuildMembersChunkData`.
 
 # 0.4.16