+++ CHANGELOG.md
@@ -17,6 +17,7 @@ through rest.
 * Added `GuildFeature.WelcomeScreenEnabled`.
 * Added `MessageModifyBuilder#allowedMentions`.
 * `Embeds` can now copy their contents over to builders by using `Embed#apply(EmbedBuilder)`.
+* Added `User#premiumType`.
 
 ## Fixes